add org basics - DAL - initialisation with default org for default admin - Org Admin page inc user org membership - handle org context in session state for the selected org

+++ source/docq/manage_user_groups.py
@@ -12,8 +12,10 @@ SQL_CREATE_USER_GROUPS_TABLE = """
 CREATE TABLE IF NOT EXISTS user_groups (
     id INTEGER PRIMARY KEY,
     name TEXT UNIQUE,
+    org_id INTEGER NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    FOREIGN KEY (org_id) REFERENCES orgs (id)
 )
 """
 
@@ -39,23 +41,31 @@ def _init() -> None:
 
 
 def list_user_groups(
+    org_id: int,
     name_match: str = None,
 ) -> List[Tuple[int, str, List[Tuple[int, str]], datetime, datetime]]:
     """List user groups.
 
     Args:
+        org_id (int): The org id.
         name_match (str, optional): The group name match. Defaults to None.
 
     Returns:
         List[Tuple[int, str, List[Tuple[int, str]], datetime, datetime]]: The list of user groups.
     """
-    log.debug("Listing user groups: %s", name_match)
+    log.debug("Listing user groups that match: '%s' for org_id: %s", name_match, org_id)
+    if not org_id:
+        raise ValueError("`org_id` cannot be None.")
+
     with closing(
         sqlite3.connect(get_sqlite_system_file(), detect_types=sqlite3.PARSE_DECLTYPES)
     ) as connection, closing(connection.cursor()) as cursor:
         user_groups = cursor.execute(
-            "SELECT id, name, created_at, updated_at FROM user_groups WHERE name LIKE ?",
-            (f"%{name_match}%" if name_match else "%",),
+            "SELECT id, name, created_at, updated_at FROM user_groups WHERE org_id = ? AND name LIKE ?",
+            (
+                org_id,
+                f"%{name_match}%" if name_match else "%",
+            ),
         ).fetchall()
 
         members = cursor.execute(
@@ -67,11 +77,12 @@ def list_user_groups(
         return [(x[0], x[1], [(y[1], y[2]) for y in members if y[0] == x[0]], x[2], x[3]) for x in user_groups]
 
 
-def create_user_group(name: str) -> bool:
+def create_user_group(name: str, org_id: int) -> bool:
     """Create a user group.
 
     Args:
         name (str): The group name.
+        org_id (int): The org id.
 
     Returns:
         bool: True if the user group is created, False otherwise.
@@ -81,8 +92,8 @@ def create_user_group(name: str) -> bool:
         sqlite3.connect(get_sqlite_system_file(), detect_types=sqlite3.PARSE_DECLTYPES)
     ) as connection, closing(connection.cursor()) as cursor:
         cursor.execute(
-            "INSERT INTO user_groups (name) VALUES (?)",
-            (name,),
+            "INSERT INTO user_groups (name, org_id) VALUES (?, ?)",
+            (name, org_id),
         )
         connection.commit()
         return True
@@ -125,11 +136,12 @@ def update_user_group(id_: int, members: List[int], name: str = None) -> bool:
         return True
 
 
-def delete_user_group(id_: int) -> bool:
+def delete_user_group(id_: int, org_id: int) -> bool:
     """Delete a user group.
 
     Args:
         id_ (int): The user group id.
+        org_id (int): The org id the group belongs to.
 
     Returns:
         bool: True if the user group is deleted, False otherwise.
@@ -138,7 +150,7 @@ def delete_user_group(id_: int) -> bool:
     with closing(
         sqlite3.connect(get_sqlite_system_file(), detect_types=sqlite3.PARSE_DECLTYPES)
     ) as connection, closing(connection.cursor()) as cursor:
-        cursor.execute("DELETE FROM user_group_members WHERE group_id = ?", (id_,))
-        cursor.execute("DELETE FROM user_groups WHERE id = ?", (id_,))
+        cursor.execute("DELETE FROM user_group_members WHERE group_id = ? ", (id_,))
+        cursor.execute("DELETE FROM user_groups WHERE id = ? AND org_id = ?", (id_, org_id))
         connection.commit()
         return True
